+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### :syringe: Fixed
+
+- [#360](https://github.com/FantasticFiasco/action-update-license-year/pull/359) Unable to import GPG private key
+
 ## [3.0.1] - 2023-07-05
 
 ### :syringe: Fixed